examples/imitation_learning_openbot/requirements.txt
tensorflow-gpu
notebook
matplotlib
pillow
aiohttp[speedups]
aiohttp-devtools
aiohttp_json_rpc
aiozeroconf
imageio
aiohttp_json_rpc
openbot_frontend==0.6.0
numpyencoder
black[jupyter]
